Add files via upload
In this notebook, we will be exploring the basics of data processing and transformation using JavaScript and D3. Often when creating visualizations, you will be working with tabular data as your input. This data can be transformed in order to derive the information that you want to visualize. Typical transformations include selecting a subset of features, creating new features, filtering, sorting, and aggregating.